refactor: 헤더 수정 및 반영
